+++ commerce/src/main/java/store/_0982/commerce/domain/sellerbalance/SellerBalanceRepository.java
@@ -1,4 +1,4 @@
-package store._0982.commerce.domain.settlement;
+package store._0982.commerce.domain.sellerbalance;
 
 import java.util.Optional;
 import java.util.UUID;
