--- commerce/src/main/java/store/_0982/commerce/domain/settlement/BalanceHistoryStatus.java
@@ -1,6 +1,0 @@
-package store._0982.commerce.domain.settlement;
-
-public enum BalanceHistoryStatus {
-    CREDIT,     // 입금/증가
-    DEBIT       // 출금/감소
-}
