+++ commerce/src/main/java/store/_0982/commerce/infrastructure/settlement/SellerBalanceHistoryJpaRepository.java
@@ -3,7 +3,7 @@ package store._0982.commerce.infrastructure.settlement;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.jpa.repository.JpaRepository;
-import store._0982.commerce.domain.settlement.SellerBalanceHistory;
+import store._0982.commerce.domain.sellerbalance.SellerBalanceHistory;
 
 import java.util.UUID;
